requirements.txt
requests
gtts
pocket
pymongo
telepot==8.3
newspaper3k
pyyaml